First pass of complete scanner/decoder, with habitat upload support.
