+++ auto_rx/autorx/__init__.py
@@ -5,4 +5,4 @@
 #   Copyright (C) 2018  Mark Jessop <vk5qi@rfhead.net>
 #   Released under GNU GPL v3 or later
 #
-__version__ = "20180525"
+__version__ = "20180525-alpha"
